Add 'add result' endpoint and fix the result DTOs

+++ src/main/java/ie/shannen/runningrace/repository/model/ResultEntity.java
@@ -21,5 +21,5 @@ public class ResultEntity {
     @OneToOne
     private RaceEntity raceEntity;
 
-    private Long time;
+    private Double time;
 }
